+++ pkg/cloudconformity/conformity.go
@@ -6,9 +6,10 @@ import (
 	"fmt"
 	"io"
 	"io/ioutil"
-	"log"
 	"net/http"
 	"net/url"
+	"reflect"
+	"strings"
 )
 
 type method interface {
@@ -58,7 +59,11 @@ func newRequest(c *Client, methodType string, path string, payload io.Reader, ra
 	u.Path = resource
 	urlString := u.String()
 	client := c.HttpClient
-
+    log_debug("Request URL: %v"+urlString)
+    buf := new(strings.Builder)
+    payload_string, err := io.Copy(buf, payload)
+    log_encrypted(string(payload_string))
+    result_name := reflect.Indirect(reflect.ValueOf(result)).Type().Name()
 	req, err := http.NewRequest(methodType, urlString, payload)
 	if err != nil {
 		return nil, err
@@ -79,15 +84,16 @@ func newRequest(c *Client, methodType string, path string, payload io.Reader, ra
 	if err != nil {
 		return nil, err
 	}
-
+	log_debug("Response Body of " + result_name)
+	log_encrypted(string(body))
 	err = json.Unmarshal(body, &result)
 	if err != nil {
 		return nil, err
 	}
 
 	if resp.StatusCode != 200 {
-		log.Printf("[DEBUG] Conformity request error: %v\n", resp.StatusCode)
-		log.Printf("[DEBUG] Conformity reponse body error: %v\n", string(body))
+		log_debug("Conformity request error: "+ string(resp.StatusCode))
+		log_debug("Conformity response body error"+string(body))
 
 		return body, errors.New(string(body))
 	}