+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.3.7
+
+* Logging the request and response payload of API
+* Improved Account create and Read request and response struct
+
 ## 0.3.6
 
 * fix documentation